Made guessing based solely on cell.options
There's no need to track previous guesses in any way other than removing
the failed guess from the possible options. But when we remove it, we
want to consider the consequences of removing that option from the cell
in the context of the whole board, so we use SudokuBoard.remove_option.